try changing data type of ints to longs to account for math when printing INT_MIN

+++ get_print_func.c
@@ -25,7 +25,7 @@ int get_print_func(char t, va_list args)
 			return (print_percent());
 		case 'i':
 		case 'd':
-			return (print_int(va_arg(args, int), count_ptr));
+			return (print_int(va_arg(args, long), count_ptr));
 		default:
 			return (_putchar('%') + _putchar(t));
 	}
